Backup utility
Allows the user to choose which files are backed up with a tree-based
UI, complete with note-taking abilities.

Selected files are copied to the output directory, along with a few
generated files:
* FILES.csv - listing of all files copied with attributes (size, etc.)
* README.txt - other data about the copy operation, including notes
added in the selection step

Job data can be saved to disk for later use.  Loading job data is often
faster than re-scanning the directory.